feat(ui): a Stepper, and the hooks to drive one from a caller's own chrome
A wizard is one index and a set of steps, and every screen that has needed one so
far has rebuilt it: the manual-add flow in the torrents module carries its own
step bar, its own reachability rule and its own footer.

The parts are the kit's usual shape. What earns the component its place is the
rest:

- `useStepper()` inside a Root hands back the whole flow — where it is, how many,
  whether it may move, and the moves themselves — so a caller with its own footer
  writes one line instead of lifting the state. `<Stepper.Previous>` and
  `<Stepper.Next>` are thin over it rather than the only way to navigate.
- `useStepperItem(value)` is one step's own state, for a caller drawing its own
  indicator.
- `useSteps()` is the machine with no context and no chrome.

Reachability is a high-water mark the Root keeps, so the common flow costs the
caller nothing and a step ahead of the furthest reached leaves the D-pad ring
entirely. `complete={[...]}` is the escape for a flow that validates.

A hidden panel unmounts by default, and `keepMounted` holds it through
`<Activity mode="hidden">` rather than `display: none`: a panel merely hidden
keeps its `<Focusable>`s registered with the spatial navigator, so a television
would still stop on controls nobody can see.
